+++ webpack.dev.js
@@ -4,11 +4,12 @@ const common = require('./webpack.common.js');
 const { stylePaths } = require('./stylePaths');
 const { DefinePlugin } = require('webpack');
 
-const HOST = process.env.HOST || 'localhost';
+const HOST = process.env.HOST || '0.0.0.0';
 const PORT = process.env.PORT || '8000';
 
 const API_HOST = process.env.LISNTOR_API_HOST || 'http://192.168.123.115:3370';
 const GATEWAY_API_HOST = process.env.GATEWAY_API_HOST || 'http://192.168.123.115:8080';
+const VSAN_API_HOST = process.env.VSAN_API_HOST || 'https://10.43.4.129';
 
 module.exports = merge(common('development'), {
   mode: 'development',
@@ -31,6 +32,12 @@ module.exports = merge(common('development'), {
       },
       '/api/v2': {
         target: GATEWAY_API_HOST,
+        changeOrigin: true,
+      },
+      '/api/frontend/v1': {
+        target: VSAN_API_HOST,
+        secure: false,
+        changeOrigin: true,
       },
     },
   },